requirements.txt
solana
solders
python-dotenv
# Add any additional dependencies below as needed 
python-socketio[client]==5.8.0
colorama==0.4.6
typing-extensions==4.9.0
python-engineio==4.8.0
websockets==12.0
aiohttp==3.9.3 